fixes login and refactors some of the work done prior
