+++ src/commands/function/publish.ts
@@ -7,7 +7,7 @@ import { IDeploymentOptions } from "./interfaces";
 import { run as runBuild } from "./build";
 import { createWasmArchive, getBuildDir } from "./shared";
 import { basename, resolve } from "path";
-import { getWASMRepoServer } from "../../lib/utils";
+import { getWASMRepoServer } from "../../lib/urls";
 
 const deploymentOptions: IDeploymentOptions = {
   functionId: "",
