+++ src/store/constants.ts
@@ -1,0 +1,3 @@
+export const NAME = "bls";
+export const CONSOLE_HOST_LOCAL = "http://localhost:3005";
+export const CONSOLE_HOST_REMOTE = "https://console.bls.dev";
